requirements.txt
langchain
langchain-core
langchain-community
langchain_huggingface
langgraph
langgraph-supervisor
openai==1.35.3
google-generativeai==0.7.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.8.1
spacy==3.7.4
chromadb==0.4.24
pypdf==4.2.0
beautifulsoup4==4.12.3
elevenlabs
pillow
langchain-ollama
sentence-transformers
sounddevice
openai-whisper
PyJWT
flask-socketio
eventlet
